player progress and stage select ui rework (adding stage rating)

+++ Assets/Scripts/UI/InGame/StageInfoUI.cs
@@ -26,7 +26,7 @@ public class StageInfoUI : UICanvas
         ReduceLifePointTxt(e.Reduced);
     }
 
-    private void ReduceLifePointTxt(int value)
+    private void ReduceLifePointTxt(float value)
     {
         if (value == 0)
         {
